The table dropper AZ-9000. If que too long drop it
The que is only for render walkers really, right? Right?
I am at loss why it takes this much time now...
Probably the is_this_a_nuke_soil test x4 that is meant to shortcut things is actually pretty expensive...
The render side is pretty lightweight, or it should be so.