Fix serious error
Error was in --Friendly Fire History -> local function on_entity_died(event)

There is a unique case, when an item dies and is not within blacklisted_types (belts and so on), but is within the whitelist_types array (I would guess this is for trees?), the friendy_fire history is incremented, but not checked for a limit.

Previously, this was a none-issue, as the list would be cleared when any other case is trigger. Now, however, only 1 log entry is deleted when it overflows, and only without that case. So every time somebody triggers that special case, by lets say destroying a tree, the array is increased by one without removing the first entry.

This should hopefully fix that.