Mtn v3 - fixes
Fixed an issue whenever a player was teleported backwards when reaching the "wall" when being too far away the locomotive/collapse.
Fixed an issue when a player mined and got a reward, they get stuck in the newly placed entity.
Fixed an issue that artilleries could set their cooldown before even firing a round.
Fixed that loot chests could spawn inside a market making them unobtainable.
Adjusted most nth_tick function handlers.
Fixed an issue where wagons that spawned in the wild could be destroyed by biters/artilleries.
Reduced the "breach zone" objective to be more aligned with the rest of the other objectives.
Fixed an issue whenever a player enabled a new spell to the "mini-spell-gui" it didn't update the main slotted spell.