II: tweaks and changes
Some techs are now locked behind islands.
Text is printed out to chat when a new tech is unlocked.

Basic loot is added to the ammo chest whenever the player progresses manually towards the next island.

Fixed an issue with biters attacking an island too soon.

Fixed an issue with terrain generation.